+++ src/app/app-routing.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { AppComponent } from './app.component';
 import { CartListComponent } from './cart/components/cart-list/cart-list.component';
 import { ProductListComponent } from './products/components/product-list/product-list.component';
 
 const routes: Routes = [
   { path: 'cart', component: CartListComponent },
-  { path: '', component: ProductListComponent, pathMatch: 'full' }
+  { path: 'products', component: ProductListComponent },
+  { path: '', redirectTo: "products", pathMatch: 'full' }
 ];
 
 @NgModule({