+++ src/app/cart/components/cart-list/cart-list.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, DoCheck, OnInit } from '@angular/core';
 import { CartItemModel } from '../../models/cart-item.model';
 import { CartService } from '../../services/cart.service';
 
@@ -7,15 +7,33 @@ import { CartService } from '../../services/cart.service';
   templateUrl: './cart-list.component.html',
   styleUrls: ['./cart-list.component.scss']
 })
-export class CartListComponent implements OnInit {
+export class CartListComponent implements OnInit, DoCheck {
 
-  items: CartItemModel[] = [];
+  items: ReadonlyArray<CartItemModel> = [];
 
-  constructor() {
-    this.items = new CartService().getCartItems();
-   }
+  constructor(private readonly cartService: CartService) { }
 
-  ngOnInit(): void {
+  onItemQuantityDecreased(item: CartItemModel): void {
+    this.cartService.decreaseQuantityByOne(item.product);
   }
 
+  onItemQuantityIncreased(item: CartItemModel): void {
+    this.cartService.increaseQuantityByOne(item.product);
+  }
+
+  onItemRemoved(item: CartItemModel): void {
+    this.cartService.removeProductFromCart(item.product);
+  }
+
+  getCartTotalPrice(): number {
+    return this.cartService.getCartTotalPrice();
+  }
+
+  ngOnInit(): void {
+    this.items = this.cartService.getCartItems();
+  }
+
+  ngDoCheck(): void {
+    this.items = this.cartService.getCartItems();
+  }
 }