+++ src/app/cart/models/cart-item.model.ts
@@ -1,4 +1,6 @@
-export class CartItemModel{
-    name: string;
-    imageUrl: string;
+import { ProductModel } from 'src/app/products/models/product.model';
+
+export interface CartItemModel {
+    readonly product: ProductModel;
+    quantity: number;
 }