+++ src/app/cart/services/cart.service.ts
@@ -1,26 +1,70 @@
+import { Injectable } from '@angular/core';
+
 import { ProductModel } from 'src/app/products/models/product.model';
-import { ProductsService } from 'src/app/products/services/products.service';
-import { CartItemComponent } from '../components/cart-item/cart-item.component';
 import { CartItemModel } from '../models/cart-item.model';
 
+@Injectable({
+    providedIn: 'root'
+})
 export class CartService {
 
-    private productsService: ProductsService;
+    private cartItems: CartItemModel[] = [];
 
-    constructor() {
-        this.productsService = new ProductsService();
+    getCartItems(): ReadonlyArray<CartItemModel> {
+        return this.cartItems;
     }
 
-    getCartItems(): CartItemModel[] {
-        return this.productsService
-            .getProducts()
-            .map(this.toCartItem);
+    getCartItemTotalPrice(cartItem: CartItemModel): number {
+        return cartItem.product.price * cartItem.quantity;
     }
 
-    private toCartItem(source: ProductModel): CartItemModel {
+    getCartTotalPrice(): number {
+        return this.cartItems.map(this.getCartItemTotalPrice)
+            .reduce((prev, next) => prev + next, 0);
+    }
+
+    addProductToCart(product: ProductModel): void {
+        const existingItem = this.searchCartForProductByName(product.name);
+        if (existingItem !== undefined) {
+            this.increaseQuantityByOne(product);
+        } else {
+            this.cartItems.push(this.createNewCartItem(product));
+        }
+    }
+
+    increaseQuantityByOne(product: ProductModel): void {
+        const itemToIncrease = this.searchCartForProductByName(product.name);
+        if (itemToIncrease !== undefined) {
+            itemToIncrease.quantity++;
+        }
+    }
+
+    decreaseQuantityByOne(product: ProductModel): void {
+        const itemToDecrease = this.searchCartForProductByName(product.name);
+        if (itemToDecrease === undefined) {
+            return;
+        }
+
+        if (itemToDecrease.quantity === 1) {
+            this.removeProductFromCart(product);
+        } else {
+            itemToDecrease.quantity--;
+        }
+    }
+
+    removeProductFromCart(product: ProductModel): void {
+        const productToRemove = this.searchCartForProductByName(product.name);
+        this.cartItems = this.cartItems.filter(item => item !== productToRemove);
+    }
+
+    private searchCartForProductByName(productName: string): CartItemModel {
+        return this.cartItems.find(item => item.product.name === productName);
+    }
+
+    private createNewCartItem(product: ProductModel): CartItemModel {
         return {
-            name: source?.name,
-            imageUrl: source?.imageUrl
+            product,
+            quantity: 1
         };
     }
 }