+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,4 +1,6 @@
 import { Component, OnInit } from '@angular/core';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { CartService } from 'src/app/cart/services/cart.service';
 
 import { ProductModel } from '../../models/product.model';
 import { ProductsService } from '../../services/products.service';
@@ -12,11 +14,20 @@ export class ProductListComponent implements OnInit {
 
   products: ProductModel[] = [];
 
-  constructor() {
-    this.products = new ProductsService().getProducts();
+  constructor(private readonly productsService: ProductsService,
+    private readonly cartService: CartService,
+    private readonly snackBar: MatSnackBar) { }
+
+  onAddedToCart(product: ProductModel): void {
+    this.cartService.addProductToCart(product);
+    this.snackBar.open(`Product ${product.name} was added to cart`, null, {
+      duration: 2000,
+      verticalPosition: 'bottom',
+      horizontalPosition: 'end'
+    });
   }
 
   ngOnInit(): void {
+    this.products = this.productsService.getProducts();
   }
-
 }